Multiple Changes
Added Preferences for Pebble special value and text (to replace BAZINGA!).
Added automation to change of units so that High and Low values and Pebble
Special Value are
altered appropriate to the units selected.
Added automatically setting Broadcast Locally on if Send to Pebble is
turned on.
Added automatically setting Send to Pebble off if Broadcast Locally is
turned off.